requirements.txt
# Core web framework - pinned for Python 3.11
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.4.0,<3.0.0

# Storage backends (AWS S3 or Supabase)
boto3==1.36.15
botocore==1.36.15
supabase==2.9.1

# Audio processing 
pydub==0.25.1
pyloudnorm==0.1.1
midiutil==1.2.1
pretty_midi==0.2.10
mido==1.3.2

# Data processing
numpy==2.3.2
pandas==2.2.3

# OpenAI integration
openai==1.58.1
groq>=0.7.0

# API and serialization
python-multipart==0.0.20
PyYAML==6.0.2
python-json-logger>=2.0.0

# Core dependencies
requests==2.32.3
python-dateutil==2.9.0.post0

# Testing dependencies
pytest==8.4.1
pytest-asyncio==0.25.0
pytest-cov==6.2.1
moto[s3]>=4.2.0
httpx>=0.25.0

# Development dependencies (optional)
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
